refactoring because of update on Cake.Bridge.DependencyInjection.Testing NuGet
