+++ src/AZDOI.Tests/Services/TestMarkdownService.cs
@@ -1,10 +1,9 @@
 ﻿using AZDOI.Services;
-using Cake.Core;
 using Microsoft.Extensions.Time.Testing;
 
 namespace AZDOI.Tests.Services;
 
-public class TestMarkdownService(ICakeContext cakeContext, FakeTimeProvider fakeTimeProvider) 
+public class TestMarkdownService(ICakeContext cakeContext, FakeTimeProvider fakeTimeProvider)
     : MarkdownServiceBase<string>(cakeContext, fakeTimeProvider)
 {
     protected override async Task WriteIndex(FileTextWriter writer, string value)
