+++ src/AZDOI.Tests/Unit/Markdown/Repositories/RepositoriesMarkdownServiceTests.cs
@@ -1,6 +1,6 @@
 ﻿using AZDOI.Services.Markdown.Repositories;
 
-namespace AZDOI.Tests.Unit.Markdown;
+namespace AZDOI.Tests.Unit.Markdown.Repositories;
 
 public class RepositoriesMarkdownServiceTests
 {
