+++ src/Directory.Packages.props
@@ -6,12 +6,13 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageVersion Include="Azure.Identity" Version="1.14.0" />
-    <PackageVersion Include="Cake.Bridge.DependencyInjection" Version="2025.5.15.133" />
+    <PackageVersion Include="Cake.Bridge.DependencyInjection" Version="2025.5.16.136" />
+    <PackageVersion Include="Cake.Bridge.DependencyInjection.Testing" Version="2025.5.16.136" />
     <PackageVersion Include="Cake.Common" Version="5.0.0" />
     <PackageVersion Include="Cake.Testing" Version="5.0.0" />
     <PackageVersion Include="coverlet.collector" Version="6.0.4" />
     <PackageVersion Include="Devlead.Console" Version="2025.5.14.181" />
-    <PackageVersion Include="Devlead.Testing.MockHttp" Version="2025.4.8.115" />
+    <PackageVersion Include="Devlead.Testing.MockHttp" Version="2025.5.14.163" />
     <PackageVersion Include="Microsoft.Extensions.Http" Version="9.0.5" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
     <PackageVersion Include="NSubstitute" Version="5.3.0" />
